pyproject.toml
[tool.poetry]
name = "mytown"
version = "0.1.0"
description = ""
authors = ["Aniruddha Adhikary <aniruddha@adhikary.net>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
django-fsm = "^2.8.0"
Pillow = "^9.2.0"
django-bulma = "^0.8.3"
djangorestframework = "^3.13.1"
Markdown = "^3.4.1"
django-crispy-forms = "^1.14.0"
crispy-bulma = "^0.8.0"
django-allauth = "^0.51.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
